Velocity Truck babyy
Used the environment gen to create a level for the brand new velocity truck! Also started implementing a better way to spawn underground mesh, along with creating a bool for the bezier curves that handles the generation update s that bezier curves can be custom made. See starting hill object in the DriveTest scene