superior paw movement
Paw movement has been coded so as to somewhat resemble the intended movement, while following the guidelines in comments as close as possible. I may have messed up as some point, if so my bad sorry.

Also! Turns out Lerp slows down as it gets closer, meaning it's never 'at' the location it needs to be. Solved this by adding a 'approximateProximity' public variable, which determines how close the player must be to the point to be considered 'touching'. Also a function but that only matters to coders and as we all know coders are the only people in the club that never use github