Updated hand movement
Modified hand movement to match new direction desired by design. It now performs the whole movement after a single button press, no submittal required. Additionally, the beginning of the ingredient being moved by the hand has been implemented, by which I mean the ingredient gets added when the hand reaches the taco. It's marked as held by the hand in the code until it places it down, but currently this is not shown visually.