Made Flip Boost Proportional to Flip Count
Added code into FlipTracker.cs and Vehicle.cs that allows the flip boost to scale off of the number of flips done. Both the boost effect and the amount of time the boost takes place are editable values in the Unity editor. Also edited the VelocityTruck Prefab to allow for the scaling of the boost sprite and the editable values.