Connected the Nitro in Taco Making to the Nitro in Driving
Updated GameManager.cs to store the number of nitro charges from tacoGameManager at the end of Taco Making and then at the start of Driving, pass that numebr to drivingGameManager. Then, in DrivingGameManager.cs, initialized the number of nitro charges of the vehicle to the given number and updated the driving UI accordingly.