Loaded Next Scene in the Background at the End of Driving
In GameManager.cs created a function to load a scene in the background of the current scene. Modified LoadTacoMakingScene() and the success end of driving to use this function. DrivingGameManager.cs: Added editor variables for the success text and fail text. Modified DrivingUIManager.cs to use the new background loading.