Set up "final" branch
Set up the new final branch, although this is not "final". This is just a branch off of one of the last masters (not the final master), consisting of some camerahandler stuff from an older branch (none of which is final either) to complete it on a more recent branch (which is also not final). final final final final final. first push, no new stuff, will add to/finish soon