Gave Taco Making a Tutorial
Made a tutorial popup at the start of taco making that is dismissible with SPACE. Tutorial can only happen in the first taco making level and can be turned on and off using the activeTutorial variable in TacoMakingGameManager.cs. Current known bug: when exiting to the menu (or other scene) while the tutorial is on screen, time scale stays at 0 instead of resetting to 1.